Move cli.py to _internal
