+++ certbot/_internal/client.py
@@ -20,7 +20,7 @@ from acme.magic_typing import Optional, List  # pylint: disable=unused-import,no
 import certbot
 from certbot._internal import account
 from certbot._internal import auth_handler
-from certbot import cli
+from certbot._internal import cli
 from certbot._internal import constants
 from certbot import crypto_util
 from certbot._internal import eff
