+++ certbot/_internal/renewal.py
@@ -15,7 +15,7 @@ import zope.component
 
 from acme.magic_typing import List  # pylint: disable=unused-import, no-name-in-module
 
-from certbot import cli
+from certbot._internal import cli
 from certbot import crypto_util
 from certbot import errors
 from certbot._internal import hooks
