+++ certbot/tests/client_test.py
@@ -462,7 +462,7 @@ class ClientTest(ClientTestCommon):
         names = [call[0][0] for call in mock_storage.call_args_list]
         self.assertEqual(names, ["example_cert", "example.com", "example.com"])
 
-    @mock.patch("certbot.cli.helpful_parser")
+    @mock.patch("certbot._internal.cli.helpful_parser")
     def test_save_certificate(self, mock_parser):
         # pylint: disable=too-many-locals
         certs = ["cert_512.pem", "cert-san_512.pem"]
