--- docs/api/cli.rst
@@ -1,5 +1,0 @@
-:mod:`certbot.cli`
-----------------------
-
-.. automodule:: certbot.cli
-   :members:
